refactor(loader): split graph-validation + graph-construction out of loader.ts
loader.ts was 576 lines running three distinct pipeline phases in one
file: YAML parse + Zod schema, pre-build validation (return schemas +
expressions + enum literal compliance), and graphlib construction +
topology checks. The review flagged it as doing too much; the phases
are genuinely separable and each has its own identity.

Three-way internal split, public API preserved so every existing
caller (graph-resolution, cli/*, engine/*, tools/*, watcher,
builder, core/index) continues to `import { ... } from "./loader.js"`
unchanged:

- **src/graph-validation.ts** (new, 155 lines) — pure pre-build
  validation: validateReturnSchemas, validateExpressions, plus two
  helpers (extractContextEnums, checkEnumCompliance) that are
  internal to the module. Runs after Zod parsing, before graphlib
  construction. No I/O, no graphlib imports, no side effects.

- **src/graph-construction.ts** (new, 135 lines) — graphlib Graph
  construction and topology checks: buildAndValidateGraph +
  validateCycles. Owns the @dagrejs/graphlib createRequire dance in
  its own copy (the CJS/ESM interop comment is now local to the
  file that uses it). Enforces startNode existence, edge target
  resolution, terminal rules, gate/wait requirements, reachability,
  and cycle breaking nodes.

- **src/loader.ts** (shrunk 576 → 327) — public API: the multi-file
  loaders (loadGraphs, loadGraphsCollecting, loadGraphsLayered),
  findGraphFiles, loadSingleGraph, validateAndBuild (now a 3-line
  orchestrator calling the two new modules), resolveContextDefaults,
  and validateCrossGraphRefs (kept here because it's only called
  from the multi-file loaders in this same file). loader.ts no
  longer needs the runtime `Graph` construct — it only imports the
  type, so the createRequire block moves out.

No behavior change, no external API change, 576/576 tests pass,
Biome clean. The file is still the entry point; it just orchestrates
instead of doing every phase itself.
